requirements.txt
Flask==3.1.2
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.7.2
mysql-connector-python==9.5.0
matplotlib==3.10.7
python-dotenv==1.2.1
openpyxl==3.10.1